Refactor UI helpers into helper module for better organization
