+++ src/bonsai/bonsai/bim/helper.py
@@ -784,3 +784,148 @@ def draw_filter(
             op.group_index = i
             op.index = j
             op.module = module
+
+# ============================================================================
+# UI Panel Visibility Helpers
+# ============================================================================
+
+def get_tab_names():
+    from bonsai.bim.prop import get_tab
+    enum_items = get_tab(None, None)
+    # Exclude None separators and the BLENDER tab (not part of BIM tab system)
+    return [item[0] for item in enum_items if item is not None and item[0] != "BLENDER"]
+
+
+def get_panel_tab_name(panel_class):
+    if hasattr(panel_class, 'bim_tab_name'):
+        return panel_class.bim_tab_name
+    return "PROJECT"  # Default fallback
+
+
+def should_show_panel(panel_id, panel_tab_name, context):
+    if tool.Blender.is_tab(context, "BOOKMARK"):
+        return is_panel_bookmarked(panel_id) and get_panel_visibility(panel_id, "BOOKMARK")
+    
+    if tool.Blender.is_tab(context, panel_tab_name):
+        return get_tab_visibility(panel_tab_name) and get_panel_visibility(panel_id, panel_tab_name)
+    
+    return False
+
+
+def get_tab_visibility(tab_name):
+    bim_props = tool.Blender.get_bim_props()
+    tab_vis = bim_props.tab_visibilities.get(tab_name)
+    return tab_vis.is_visible if tab_vis else True
+
+
+def set_tab_visibility(tab_name, visible):
+    bim_props = tool.Blender.get_bim_props()
+    tab_vis = bim_props.tab_visibilities.get(tab_name)
+    if tab_vis:
+        tab_vis.is_visible = visible
+    else:
+        new_tab = bim_props.tab_visibilities.add()
+        new_tab.name = tab_name
+        new_tab.is_visible = visible
+
+
+def get_panel_visibility(panel_id, current_tab=None):
+    panel_config = get_panel_config(panel_id)
+    if panel_config:
+        if current_tab == "BOOKMARK":
+            return panel_config.is_visible_in_bookmarks
+        else:
+            return panel_config.is_visible_in_tab
+    return True
+
+
+def is_panel_bookmarked(panel_id):
+    panel_config = get_panel_config(panel_id)
+    if panel_config:
+        return panel_config.is_bookmarked
+    return False
+
+
+def get_panel_config(panel_id, create_if_missing=False):
+    try:
+        bim_props = tool.Blender.get_bim_props()
+    except (AttributeError, AssertionError):
+        return None
+    
+    for prop in bim_props.panel_properties:
+        if prop.name == panel_id:
+            return prop
+    
+    if create_if_missing:
+        try:
+            prop = bim_props.panel_properties.add()
+            prop.name = panel_id
+            prop.is_visible_in_tab = True
+            prop.is_visible_in_bookmarks = True
+            prop.is_bookmarked = False
+            return prop
+        except AttributeError:
+            pass
+    
+    return None
+
+
+
+def get_all_tab_panels(force_refresh=False):
+    panels = {tab_name: [] for tab_name in get_tab_names() if tab_name != "BOOKMARK"}
+    panels["BOOKMARK"] = []
+    
+
+    bim_props = tool.Blender.get_bim_props()
+    for prop in bim_props.panel_properties:
+        panel_class = getattr(bpy.types, prop.name, None)
+        if panel_class:
+            tab_name = get_panel_tab_name(panel_class)
+            if tab_name and tab_name != "BOOKMARK":
+                bl_label = getattr(panel_class, "bl_label", prop.name)
+                panels[tab_name].append({"bl_idname": prop.name, "bl_label": bl_label})
+        
+        if prop.is_bookmarked:
+            panel_class = getattr(bpy.types, prop.name, None)
+            if panel_class:
+                bl_label = getattr(panel_class, "bl_label", prop.name)
+                panels["BOOKMARK"].append({"bl_idname": prop.name, "bl_label": bl_label})
+    
+    if not panels["BOOKMARK"]:
+        panels["BOOKMARK"] = [{}]
+    
+    return panels
+
+
+def initialize_tab_visibilities():
+    bim_props = tool.Blender.get_bim_props()
+    
+    if len(bim_props.tab_visibilities) > 0:
+        return
+    
+    for tab_name in get_tab_names():
+        tab_vis = bim_props.tab_visibilities.add()
+        tab_vis.name = tab_name
+        tab_vis.is_visible = True
+
+
+def initialize_panel_properties():
+    
+    bim_props = tool.Blender.get_bim_props()
+    
+    if len(bim_props.panel_properties) > 0:
+        return
+    
+    for attr_name in dir(bpy.types):
+        if attr_name.startswith("BIM_PT_tab_"):
+            panel_class = getattr(bpy.types, attr_name)
+            if not hasattr(panel_class, 'bl_idname'):
+                continue
+            
+            panel_id = panel_class.bl_idname
+            
+            prop = bim_props.panel_properties.add()
+            prop.name = panel_id
+            prop.is_visible_in_tab = True
+            prop.is_visible_in_bookmarks = True
+            prop.is_bookmarked = False
